Separating test scenarios of Endpoints, Credentials, Settings
