+++ functional/cypress-tests/cypress/support/credentialsEndpoints.js
@@ -26,6 +26,18 @@ Cypress.Commands.add('openSettings', selectedTab => {
   }
 });
 
+Cypress.Commands.add('deleteAll', (prefix, deleteButtonId) => {
+  cy.get('li.MuiListItem-container').each($li => {
+    const label = $li
+      .find('.MuiListItem-root .MuiListItemText-root span')
+      .text();
+    if (label.indexOf(prefix) === 0) {
+      cy.wrap($li.find(`[data-cy="${deleteButtonId}"]`)).click();
+      cy.get('[data-cy="confirmation-dialog-ok"]').click();
+    }
+  });
+});
+
 Cypress.Commands.add('setTestEndpoints', (testEndpoints, authToken) => {
   cy.request({
     method: 'POST',
